Update CESP-subway-proto-aboutplan.html
•	Paragraph “when you open a plan, you may be eligible for CLB…etc”
o	Change to “When you open an RESP, you may be eligible for the Bond and grants.” 
o	Add in-text link to bring to second H2 – The Bond and grants you could get. 
o	Merge this line of text with first paragraph 
•	Move up “anyone can open” sentence 
•	Remove bold from “special savings account”
•	Remove tooltip definition for RESP everywhere
•	Expand collapse for definition: fiddle with spacing to make prettier
•	Under The Bond and grants you could get
o	Add period to end of Grant bullet
o	There are provincial grants that beneficiaries living in Saskatchewan, British Columbia, and Quebec may be eligible for. 
•	Under We Send out letters
o	Change wording from “benefits” to “people who are eligible for the Bond and the Grant” 
o	“if you received a letter like this”